Решение домашнего задания. Урок 5

+++ Lessons-2/MetricsAgent/Mapper/MapperProfile.cs
@@ -12,27 +12,27 @@ public class MapperProfile : Profile
         CreateMap<CpuMetric, CpuMetricDto>();
         CreateMap<CpuMetricCreateRequest, CpuMetric>()
             .ForMember(x => x.Value, opt => opt.MapFrom(src => src.Value))
-            .ForMember(x => x.Time, opt => opt.MapFrom(src => (int)src.Time.TotalSeconds));
+            .ForMember(x => x.Time, opt => opt.MapFrom(src => (long)src.Time.TotalSeconds));
 
         CreateMap<DotnetMetric, DotnetMetricDto>();
         CreateMap<DotnetMetricCreateRequest, DotnetMetric>()
             .ForMember(x => x.Value, opt => opt.MapFrom(src => src.Value))
-            .ForMember(x => x.Time, opt => opt.MapFrom(src => (int)src.Time.TotalSeconds));
+            .ForMember(x => x.Time, opt => opt.MapFrom(src => (long)src.Time.TotalSeconds));
 
         CreateMap<HddMetric, HddMetricDto>();
         CreateMap<HddMetricCreateRequest, HddMetric>()
             .ForMember(x => x.Value, opt => opt.MapFrom(src => src.Value))
-            .ForMember(x => x.Time, opt => opt.MapFrom(src => (int)src.Time.TotalSeconds));
+            .ForMember(x => x.Time, opt => opt.MapFrom(src => (long)src.Time.TotalSeconds));
 
         CreateMap<NetworkMetric, NetworkMetricDto>();
         CreateMap<NetworkMetricCreateRequest, NetworkMetric>()
             .ForMember(x => x.Value, opt => opt.MapFrom(src => src.Value))
-            .ForMember(x => x.Time, opt => opt.MapFrom(src => (int)src.Time.TotalSeconds));
+            .ForMember(x => x.Time, opt => opt.MapFrom(src => (long)src.Time.TotalSeconds));
 
         CreateMap<RamMetric, RamMetricDto>();
         CreateMap<RamMetricCreateRequest, RamMetric>()
             .ForMember(x => x.Value, opt => opt.MapFrom(src => src.Value))
-            .ForMember(x => x.Time, opt => opt.MapFrom(src => (int)src.Time.TotalSeconds));
+            .ForMember(x => x.Time, opt => opt.MapFrom(src => (long)src.Time.TotalSeconds));
 
     }
 }
