+++ Lessons-2/MetricsAgent/Models/Dto/CpuMetricDto.cs
@@ -3,5 +3,5 @@
 public class CpuMetricDto
 {
     public int Value { get; set; }
-    public int Time { get; set; }
+    public long Time { get; set; }
 }
