+++ Lessons-2/MetricsAgent/Models/RamMetric.cs
@@ -4,5 +4,5 @@ public class RamMetric
 {
     public int Id { get; set; }
     public int Value { get; set; }
-    public int Time { get; set; }
+    public long Time { get; set; }
 }
